+++ Modules/Explorer.ps1
@@ -1,5 +1,5 @@
 ﻿<#
- COMPARTDISK 1.4.6 - Explorer.ps1
+ COMPARTDISK 1.4.7 - Explorer.ps1
  Desenvolvido por Edsilas
  Acoes: Restart | ClearCache | Spooler | ResetView
 #>